Start of NDEF tag
