+++ main/CMakeLists.txt
@@ -6,6 +6,7 @@ idf_component_register(
         "persistent_storage/persistent_storage.c"
         "spi/spi.c"
         "rfid/rfid.c"
+        "ndef/ndef.c"
     INCLUDE_DIRS
         "."
         "wifi"
@@ -13,6 +14,7 @@ idf_component_register(
         "persistent_storage"
         "spi"
         "rfid"
+        "ndef"
 )
 
 # In release, comment out the below line
