+++ main/main.c
@@ -7,6 +7,7 @@
 #include "rfid/rfid_pcd_register_types.h"
 #include "spi/spi.h"
 // #include "wifi/wifi.h"
+#include "ndef/ndef.h"
 
 #include <driver/spi_common.h>
 #include <driver/spi_master.h>
@@ -32,68 +33,6 @@ esp_err_t setup(spi_device_handle_t* rfid_spi_handle) {
 	return ESP_OK;
 }
 
-char represent_byte(uint8_t byte) {
-	if ((byte >= '0' && byte <= '9') || (byte >= 'a' && byte <= 'z') ||
-		(byte >= 'A' && byte <= 'Z')) {
-		return (char)byte;
-	}
-	return '.';
-}
-
-void print_buffer(uint8_t* buffer, uint8_t size, uint8_t address) {
-	// Output the data
-	const uint8_t width = 4;
-	const uint8_t left_padding = 2;
-	const uint8_t right_padding = 2;
-	const uint8_t middle_left = 1;
-	const uint8_t middle_right = 2;
-	const uint8_t line_width = left_padding + width * 3 - 1 +
-							   (middle_left + middle_right) + width * 2 +
-							   right_padding;
-	const uint8_t middle_start = left_padding + width * 3 - 1;
-	const uint8_t block_data = 16;
-
-	char line[line_width];
-	for (uint8_t x = 0; x < line_width; x++) {
-		line[x] = '#';
-	}
-	line[0] = '|';
-	line[1] = ' ';
-
-	line[middle_start] = ' ';
-	line[middle_start + 1] = '|';
-	line[middle_start + 2] = ' ';
-
-	line[line_width - 3] = ' ';
-	line[line_width - 2] = '|';
-	line[line_width - 1] = '\0';
-
-	LOG("+----+-------------+---------+");
-	LOG("|ADDR| BYTES       | CHARS   |");
-	for (uint8_t y = 0; y < size; y += width) {
-		if (y % block_data == 0) {
-			LOG("+----+-------------+---------+");
-		}
-
-		for (uint8_t x = 0; x < width; x++) {
-			uint8_t index = y + x;
-			sprintf(line + 3 * x + left_padding, "%02x ", buffer[index]);
-		}
-		line[middle_start + 1] = '|';
-		for (uint8_t x = 0; x < width; x++) {
-			uint8_t index = y + x;
-			sprintf(
-				middle_start + (middle_left + middle_right) + line + 2 * x,
-				"%c ", represent_byte(buffer[index])
-			);
-		}
-		line[line_width - 2] = '|';
-		LOG("| %02x %s", address, line);
-		address += 4;
-	}
-	LOG("+----+-------------+---------+");
-}
-
 void app_main(void) {
 
 	spi_device_handle_t rfid_handle = {0};
@@ -102,14 +41,7 @@ void app_main(void) {
 	// Wake up the rfid reader
 	rfid_wakeup_mifare_tag(&rfid_handle);
 
-	// Read actual data from the tag
-	uint8_t buffer[128] = {0};
-	for (uint8_t sectors = 0; sectors < 8; sectors++) {
-		rfid_read_mifare_tag(
-			&rfid_handle, 4 + sectors * 4, buffer + sectors * 16, 16
-		);
-	}
-	print_buffer(buffer, 128, 4);
+	test(&rfid_handle);
 
 	// Read actual data from the tag
 	// rfid_read_mifare_tag(&rfid_handle, 4 + 4, buffer, 16);
